+++ source/global.d.ts
@@ -44,6 +44,26 @@ interface DictionaryProvider {
   setUserDictionary: (dict: string[]) => void
 }
 
+// Dictionary in the form dic['yyyy-mm-dd'] = value
+interface DailyDictionary {
+  [day: string]: number
+}
+
+// Statistics object
+interface Stats {
+  wordCount: DailyDictionary // All words for the graph
+  pomodoros: DailyDictionary // All pomodoros ever completed
+  avgMonth: number // Monthly average
+  today: number // Today's word count
+  sumMonth: number // Overall sum for the past month
+}
+
+interface StatsProvider {
+  increaseWordCount: (words: number) => void
+  increasePomodoros: () => void
+  getData: () => Stats
+}
+
 // Before the log provider has booted, these messages will be added to the
 // preBootLog
 interface BootLog {
@@ -75,5 +95,6 @@ declare module NodeJS {
     targets: any
     recentDocs: any
     tags: any
+    stats: StatsProvider
   }
 }